rename WeaponUtil to Abstract
