+++ include/Item/Weapon/WoodenShield.hpp
@@ -1,7 +1,7 @@
 #ifndef WOODENSHIELD_HPP
 #define WOODENSHIELD_HPP
 
-#include "Item/Weapon/WeaponUtil/Shield.hpp"
+#include "Item/Weapon/Abstract/Shield.hpp"
 
 class WoodenShield : public Shield {
 public:
@@ -9,7 +9,6 @@ public:
     ~WoodenShield() = default;
 
     void Parry(std::shared_ptr<Mob>& mob, const glm::vec2& Dir = {1, 1}) override;
-
 
 private:
     int ParryDamage = 70;
